requirements.txt
pillow
numpy 
RPi.GPIO
spidev